feat: Add AutoMapper to project for mapping data
Added a new dependency package AutoMapper.Extensions.Microsoft.DependencyInjection to the API.csproj file and added the service to ApplicationServiceExtensions.cs. Created two DTO objects MemberDto and PhotoDto to hold the data from the AppUser and Photo entities. Updated the UsersController to use IUserRepository and IMapper interface as parameters in the constructor. Modified the GetUsers and GetUser methods to fetch users from the database using IUserRepository, map the data to the newly created DTO objects using IMapper and then return the results.  Lastly, I created an AutoMapperProfiles.cs class which creates mapping between the entities and DTOs.