Add member and photo interfaces; add account and members services
Interfaces for Member and Photo were added to the project. The Member interface is used to describe a user, and the Photo interface describes a photo object. The AccountService and MembersService services were added to the client/src/app/_services directory. The AccountService service contains the logic used for authenticating a user, while the MembersService service is used to retrieve a list of users and their details. The environment file was also modified. The production environment file has been updated to include an apiUrl, and the development environment file has been updated with a url for retrieving the apiUrl.